modified more routes, with updated error messsage

+++ routes/api/product-routes.js
@@ -57,50 +57,41 @@ router.post('/', (req, res) => {
 });
 
 // update product
-router.put('/:id', (req, res) => {
-  // update product data
-  Product.update(req.body, {
-    where: {
-      id: req.params.id,
-    },
-  })
-    .then((product) => {
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .catch((err) => {
-      res.status(400).json({ message: 'Product update failed.', error: err });
-    })
-    .then((productTags) => {
-          // create filtered list of new tag_ids
-          const productTagIds = productTags.map(({ tag_id }) => tag_id);
-          
-          const newProductTags = req.body.tagIds
-            ? req.body.tagIds
-                .filter((tag_id) => !productTagIds.includes(tag_id))
-                .map((tag_id) => {
-                  return {
-                    product_id: req.params.id,
-                    tag_id,
-                  };
-                })
-            : [];
-            
+router.put('/:id', async (req, res) => {
+  try {
+    await Product.update(req.body, { where: { id: req.params.id } });
 
-            // figure out which ones to remove
-          const productTagsToRemove = productTags
-          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-          .map(({ id }) => id);
-                  // run both actions
-          return Promise.all([
-            ProductTag.destroy({ where: { id: productTagsToRemove } }),
-            ProductTag.bulkCreate(newProductTags),
-          ]);
-        })
-        .then(() => res.status(200).json({ message: 'Product updated.'}))
-        .catch((err) => {
-          res.status(400).json(err);
+    if (req.body.tags && req.body.tags.length > 0) {
+      const productTags = await ProductTag.findAll({ where: { product_id: req.params.id } });
+      const productTagIds = productTags.map(({ tag_id }) => tag_id);
+
+      const newProductTags = req.body.tags
+        .filter((tag_id) => !productTagIds.includes(tag_id))
+        .map((tag_id) => {
+          return {
+            product_id: req.params.id,
+            tag_id,
+          };
         });
-    });
+
+      const productTagsToRemove = productTags
+        .filter(({ tag_id }) => !req.body.tags.includes(tag_id))
+        .map(({ id }) => id);
+      
+      await Promise.all([
+        ProductTag.destroy({ where: { id: productTagsToRemove } }),
+        ProductTag.bulkCreate(newProductTags),
+      ]);
+    }
+
+    const product = await Product.findByPk(req.params.id, { include: [{ model: Tag }] });
+    return res.status.json(product);
+  } catch (err) {
+    console.log(error);
+    res.status(500).json({ message: 'Product update failed.' });
+  }
+});
+  
 
 router.delete('/:id', async (req, res) => {
   // delete one product by its `id` value
